requirements.txt
APScheduler==3.2.0
requests==2.10.0
slackclient==1.0.1
PyYAML==3.11
